notes
okay, obvious in hindsight, but:
- most meaningful characters are CHR chunks (or, they're CHR once and
then repeated later with REFs). so handling how expensive
they are (anywhere from 6 to 14 bits per char) is important.
and it's a major consideration when deciding whether or not to use a
RAW chunk.
- to continue down this road to 'perfect compression', you'd need to
save _every_ compression of each prefix, or at least you could only
say one was definitely better than another if they had the same exact
mtf table. so this is sorta infeasible (for large programs...)